+++ lib/blizzard_api/hearthstone/game_data/metadata.rb
@@ -6,6 +6,7 @@ module BlizzardApi
     # This class allows access to Hearthstone metadata data
     #
     # @see https://develop.battle.net/documentation/api-reference/hearthstone-game-data-api
+    # https://develop.battle.net/documentation/hearthstone/guides/metadata
     #
     # You can get an instance of this class using the default region as follows:
     #   api_instance = BlizzardApi::Hearthstone.metadata