+++ test/blizzard/wow/game_data/achievement_test.rb
@@ -8,7 +8,7 @@ module BlizzardApi
       def test_achievement_index
         achievement = BlizzardApi::Wow::Achievement.new
         achievement_data = achievement.index
-        assert_equal 5034, achievement_data[:achievements].count
+        assert_equal 5131, achievement_data[:achievements].count
       end
 
       def test_achievement_get
@@ -31,7 +31,7 @@ module BlizzardApi
         achievement = BlizzardApi::Wow::Achievement.new
         achievement_categories_data = achievement.categories
         assert_equal 14, achievement_categories_data[:root_categories].count
-        assert_equal 113, achievement_categories_data[:categories].count
+        assert_equal 115, achievement_categories_data[:categories].count
         assert_equal 24, achievement_categories_data[:guild_categories].count
       end
 
@@ -39,7 +39,7 @@ module BlizzardApi
         achievement = BlizzardApi::Wow::Achievement.new
         achievement_categories_data = achievement.category 81
         assert_equal 'Feats of Strength', achievement_categories_data[:name][:en_US]
-        assert_equal 108, achievement_categories_data[:achievements].count
+        assert_equal 109, achievement_categories_data[:achievements].count
       end
 
       def test_achievement_media