+++ test/blizzard/wow/game_data/creature_test.rb
@@ -61,7 +61,7 @@ module BlizzardApi
 
       def test_creature_display_media
         creature_display_media_data = @creature.display_media 30_221
-        assert_equal 'https://render-us.worldofwarcraft.com/npcs/portrait/creature-display-30221.jpg',
+        assert_equal 'https://render-us.worldofwarcraft.com/npcs/zoom/creature-display-30221.jpg',
                      creature_display_media_data[:assets][0][:value]
 
         creature_display_media_data = @creature.display_media 180, classic: true